// src/types/theme.ts
export enum BUTTON_TYPE {
  FILL = 'fill',
  FILLROUNDED = 'fillrounded',
  FILLROUNDEDFULL = 'fillroundedfull',
  OUTLINE = 'outline',
  OUTLINEROUNDED = 'outlinerounded',
  OUTLINEROUNDEDFULL = 'outlineroundedfull',
  SOFTSHADOW = 'softshadow',
  SOFTSHADOWROUNDED = 'softshadowrounded',
  SOFTSHADOWROUNDEDFULL = 'softshadowroundedfull',
  HARDSHADOW = 'hardshadow',
  HARDSHADOWROUNDED = 'hardshadowrounded',
  HARDSHADOWROUNDEDFULL = 'hardshadowroundedfull',
}

export enum BACKGROUND_TYPE {
  SOLID = 'solid',
  GRADIENT = 'gradient',
  CUBE = 'cube',
  WIREFRAME = 'wireframe',
  COLORFUL = 'colorful',
  POLKA = 'polka',
  HEXAGON = 'hexagon',
  MOON = 'moon',
  SPRINKLE = 'sprinkle',
  CLOUDY = 'cloudy',
  IMAGE = 'image',
  COLORED_PATTERNS = 'coloredpatterns',
  COLORED_SHAPES = 'coloredshapes',
  CONTOUR_LINE = 'contourline',
  MINERAL_BLUE = 'mineralblue',
  MINERAL_GREEN = 'mineralgreen',
  MINERAL_ORANGE = 'mineralorange',
  MINERAL_YELLOW = 'mineralyellow',
}

export const BACKGROUND_LIST = [
  BACKGROUND_TYPE.SOLID as string,
  BACKGROUND_TYPE.GRADIENT as string,
  BACKGROUND_TYPE.CUBE as string,
  BACKGROUND_TYPE.COLORFUL as string,
  BACKGROUND_TYPE.POLKA as string,
  BACKGROUND_TYPE.HEXAGON as string,
  BACKGROUND_TYPE.MOON as string,
  BACKGROUND_TYPE.SPRINKLE as string,
  BACKGROUND_TYPE.CLOUDY as string,
  BACKGROUND_TYPE.CONTOUR_LINE as string,
  BACKGROUND_TYPE.COLORED_PATTERNS as string,
  BACKGROUND_TYPE.COLORED_SHAPES as string,
  BACKGROUND_TYPE.IMAGE as string,
]